Youtube
The objective of this report is to find the most viewed YouTube video categories across eight countries. The information will be used to recommend where advertising money should be alloted. The list of categories below is pulled directly from YouTube.com. The data is from usuage for one day. It contains the most watched videos of the day in each country. The data was collected from Youtube and converted into CSV files. The CSV files were then convereted to JSON files using Python.