Testing Null Hypothesis in R
a) Null: The null hypothesis is the theory presented by the Bureau of Crime Statistics and Research of Australia stating that the mean length of imprisonment for motor-vehicle-theft offenders in Australia is 16.7 months.

Alternative: The alternative hypothesis is the theory presented by the group of researchers that the mean length of imprisonment is less than 16.7 months.

b. Determine the probability of a Type I error.

For a one-tailed test, the probability of a type I error is at most alpha. For a two-tailed test, the probability is exactly alpha. Alpha is the same as the significance level which is 0.05. So there is 5% chance of rejecting the null hypothesis when it is true.

c. Determine the probability of a Type II error.