Customer Churn Analysis in Python
This project is intended to identify the churn generation of customers in a telecommunications dataset. The telecommunication sector consists of companies that make communication accessible on a global scale, whether through the telephone or Internet connection, or via cable networks or wirelessly. This is a multi-billion dollar industry that is saturated with stiff competition. Churn is a measure of how many customers stop using a service or product, often evaluated for a specific period of time. This is one of the biggest difficulties any industry faces. Machine learning models will be used to analyze patterns, trends and predictions extracted from the data in order to identify significant decreases in the use of services and products by consumers. The goal is to precisely provide a company with methods that will help identify a decrease in consumption of their products in order to subsequently identify the problems that caused this decrease. This will in turn help the company’s marketing department explore new campaigns aimed at retaining and attracting new customers with differentiated services and new products