Data Wrangling - Admissions Project in Python
The data sources were mostly found online by searching "mock college applications". I searched for applicants for several different types of programs from universities across the United States. I changed the names of the applicants as well as the universities. For some of the files, I added embellishments such as hobbies and GPA scores. For a few of the candidates who already had STEM skills, I changed the wording to more common terms that I found in other applicant files.