Salary Increases & Home Buying Correlations in R
The goal of this report is to provide data that will help jousnalists at the The San Francisco Chronicle newspaper determine if the raise in annual salaries for San Fransisco City employees is increasing at the rate of the home purchasing prices. This report will serve as a detailed analysis of employee salaries for the city San Francisco. The data will be used by the San Fransisco Chronicle newspaper. Journalists at the publication are conducting an comparison of cost of living in the city to salary averages. They have request that the report examine two main factors that possibly affect the cost of living: 1) income and housing.