Statistical Model  with Crime Data (Python)
The purpose of this report is to provide the police department a guideline to aid in enforcement optimization. This report will examine several factors in the dataset and focus on possible patterns associated with the types of crimes listed. Five columns will be explored. The first exploration will create two new data sets from the columns "report_year". The first data set will be of all crimes committed in 2014 while the second will be of all crimes committed in 2015. This will be used to find the sum of violent crimes for each year to show whether there was a decrease or increase in crime. The second observation will be in the crimes oer capita column. This obeservation will show the top 20 cities with the highest crimes per capita in 2015. The third observation will examine if population size had a large influence on crime. A major city with a sharp increase in poulation will be chosen for this observation. The observation for crime based on population increase/decrease can be performed with any city however. For the fourth observation, I will find the top five cities with the most homicides from 2012-2015. For the fifth observation, the column agency_jurisdiction will be explored for data pertaining to one city from the column.